[#4626] Fix save button disabled and yes no option
- [#4626] Replace scorecheckbox with scoringfield

- [#4626] Add disabled at scroing field
